feature: Order 추가 기능 구현

+++ src/main/java/com/example/ReservationPurchase/order/infrastructure/entity/OrderEntity.java
@@ -31,7 +31,7 @@ public class OrderEntity {
     private ProductType productType;
 
     @Column(name = "quantity", nullable = false)
-    private Long quantity;
+    private Integer quantity;
 
     @Column(name = "member_id", nullable = false)
     private Long memberId;
